requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
anywidget==0.9.18
asttokens==3.0.0
attrs==25.3.0
biopython==1.85
blinker==1.9.0
cachetools==6.1.0
certifi==2025.7.9
charset-normalizer==3.4.2
click==8.1.8
comm==0.2.2
decorator==5.2.1
executing==2.2.0
fastapi==0.116.1
filelock==3.18.0
frozenlist==1.7.0
fsspec==2025.5.1
gitdb==4.0.12
GitPython==3.1.44
gosling==0.3.0
grpclib==0.4.8
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
iniconfig==2.1.0
ipython==9.4.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.7
jedi==0.19.2
Jinja2==3.1.6
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyterlab_widgets==3.0.15
loguru==0.7.3
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mdurl==0.1.2
modal==1.0.5
mpmath==1.3.0
multidict==6.6.3
narwhals==1.46.0
networkx==3.5
numpy==2.3.1
packaging==25.0
pandas==2.3.1
parso==0.8.4
pexpect==4.9.0
pillow==11.3.0
pluggy==1.6.0
prompt_toolkit==3.0.51
propcache==0.3.2
protobuf==6.31.1
psygnal==0.14.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==20.0.0
pydantic==2.11.7
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.19.2
pysam==0.23.3
plotly
pytest==8.4.1
pytest-asyncio==1.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
referencing==0.36.2
requests==2.32.4
rich==14.0.0
rpds-py==0.26.0
shellingham==1.5.4
sigtools==4.0.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.41
stack-data==0.6.3
starlette==0.47.1
streamlit==1.46.1
streamlit-gosling==0.1.4
sympy==1.14.0
synchronicity==0.9.16
tenacity==9.1.2
toml==0.10.2
torch==2.7.1
tornado==6.5.1
traitlets==5.14.3
typer==0.16.0
types-certifi==2021.10.8.3
types-toml==0.10.8.20240310
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
watchfiles==1.1.0
wcwidth==0.2.13
widgetsnbextension==4.0.14
yarl==1.20.1

streamlit
pandas
requests
python-dotenv 
annotated-types==0.7.0
anyio==4.9.0
anywidget==0.9.18
asttokens==3.0.0
attrs==25.3.0
biopython==1.85
blinker==1.9.0
cachetools==6.1.0
certifi==2025.7.9
charset-normalizer==3.4.2
click==8.1.8
comm==0.2.2
decorator==5.2.1
executing==2.2.0
fastapi==0.116.1
filelock==3.18.0
frozenlist==1.7.0
fsspec==2025.5.1
gitdb==4.0.12
GitPython==3.1.44
gosling==0.3.0
grpclib==0.4.8
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
iniconfig==2.1.0
ipython==9.4.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.7
jedi==0.19.2
Jinja2==3.1.6
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyterlab_widgets==3.0.15
loguru==0.7.3
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mdurl==0.1.2
modal==1.0.5
mpmath==1.3.0
multidict==6.6.3
narwhals==1.46.0
networkx==3.5
numpy==2.3.1
packaging==25.0
pandas==2.3.1
parso==0.8.4
pexpect==4.9.0
pillow==11.3.0
pluggy==1.6.0
prompt_toolkit==3.0.51
propcache==0.3.2
protobuf==6.31.1
psygnal==0.14.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==20.0.0
pydantic==2.11.7
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.19.2
pysam==0.23.3
plotly
pytest==8.4.1
pytest-asyncio==1.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
referencing==0.36.2
requests==2.32.4
rich==14.0.0
rpds-py==0.26.0
shellingham==1.5.4
sigtools==4.0.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.41
stack-data==0.6.3
starlette==0.47.1
streamlit==1.46.1
streamlit-gosling==0.1.4
sympy==1.14.0
synchronicity==0.9.16
tenacity==9.1.2
toml==0.10.2
torch==2.7.1
tornado==6.5.1
traitlets==5.14.3
typer==0.16.0
types-certifi==2021.10.8.3
types-toml==0.10.8.20240310
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
watchfiles==1.1.0
wcwidth==0.2.13
widgetsnbextension==4.0.14
yarl==1.20.1
numpy
pandas
pydantic
pysam
scikit-learn
fair-esm

